+++ AWSChimeSDKMessaging/AWSChimeSDKMessagingService.m
@@ -25,7 +25,7 @@
 #import "AWSChimeSDKMessagingResources.h"
 
 static NSString *const AWSInfoChimeSDKMessaging = @"ChimeSDKMessaging";
-NSString *const AWSChimeSDKMessagingSDKVersion = @"2.33.9";
+NSString *const AWSChimeSDKMessagingSDKVersion = @"2.33.10";
 
 
 @interface AWSChimeSDKMessagingResponseSerializer : AWSJSONResponseSerializer