+++ AWSKinesisVideoArchivedMedia/AWSKinesisVideoArchivedMediaService.m
@@ -25,7 +25,7 @@
 #import "AWSKinesisVideoArchivedMediaResources.h"
 
 static NSString *const AWSInfoKinesisVideoArchivedMedia = @"KinesisVideoArchivedMedia";
-NSString *const AWSKinesisVideoArchivedMediaSDKVersion = @"2.33.9";
+NSString *const AWSKinesisVideoArchivedMediaSDKVersion = @"2.33.10";
 
 
 @interface AWSKinesisVideoArchivedMediaResponseSerializer : AWSJSONResponseSerializer