+++ AWSKinesisVideoWebRTCStorage/AWSKinesisVideoWebRTCStorageService.m
@@ -25,7 +25,7 @@
 #import "AWSKinesisVideoWebRTCStorageResources.h"
 
 static NSString *const AWSInfoKinesisVideoWebRTCStorage = @"KinesisVideoWebRTCStorage";
-NSString *const AWSKinesisVideoWebRTCStorageSDKVersion = @"2.33.9";
+NSString *const AWSKinesisVideoWebRTCStorageSDKVersion = @"2.33.10";
 
 
 @interface AWSKinesisVideoWebRTCStorageResponseSerializer : AWSJSONResponseSerializer